requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-openai
chromadb
streamlit
markdown
pypdf
openai
tiktoken
python-dotenv
pyyaml
requests
pandas
pdf2image
matplotlib
easyocr
sqlalchemy